services/api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
chromadb==0.6.3
rank-bm25>=0.2
sentence-transformers>=3.0
pydantic-settings>=2.0
sse-starlette>=2.0
